requirements.txt
fastapi
uvicorn
scikit-learn==1.8.0
pandas==3.0.0
numpy==2.4.2
joblib
psycopg2-binary
